+++ pybamm/models/full_battery_models/lithium_ion/base_lithium_ion_model.py
@@ -174,17 +174,41 @@ class BaseModel(pybamm.BaseBatteryModel):
 
         # Lithium lost to side reactions
         # Different way of measuring LLI but should give same value
-        LLI_sei = self.variables["Loss of lithium to SEI [mol]"]
-        LLI_reactions = LLI_sei
+        n_Li_lost_sei = self.variables["Loss of lithium to SEI [mol]"]
+        n_Li_lost_reactions = n_Li_lost_sei
         if "negative electrode" in domains:
-            LLI_sei_cracks = self.variables["Loss of lithium to SEI on cracks [mol]"]
-            LLI_pl = self.variables["Loss of lithium to lithium plating [mol]"]
-            LLI_reactions += LLI_sei_cracks + LLI_pl
+            n_Li_lost_sei_cracks = self.variables[
+                "Loss of lithium to SEI on cracks [mol]"
+            ]
+            n_Li_lost_pl = self.variables["Loss of lithium to lithium plating [mol]"]
+            n_Li_lost_reactions += n_Li_lost_sei_cracks + n_Li_lost_pl
 
         self.variables.update(
             {
-                "Total lithium lost to side reactions [mol]": LLI_reactions,
-                "Total capacity lost to side reactions [A.h]": LLI_reactions
+                "Total lithium lost to side reactions [mol]": n_Li_lost_reactions,
+                "Total capacity lost to side reactions [A.h]": n_Li_lost_reactions
+                * param.F
+                / 3600,
+            }
+        )
+        # Lithium lost to loss of active material
+        n_Li_lost_LAM_n = self.variables[
+            "Loss of lithium due to loss of active material in negative electrode [mol]"
+        ]
+        n_Li_lost_LAM_p = self.variables[
+            "Loss of lithium due to loss of active material in positive electrode [mol]"
+        ]
+        n_Li_lost_LAM = n_Li_lost_LAM_n + n_Li_lost_LAM_p
+        self.variables.update({"Total lithium lost to LAM [mol]": n_Li_lost_LAM})
+
+        self.variables.update(
+            {
+                "Total lithium in system [mol]": n_Li
+                + n_Li_lost_reactions
+                + n_Li_lost_LAM,
+                "Total capacity in system [A.h]": (
+                    n_Li + n_Li_lost_reactions + n_Li_lost_LAM
+                )
                 * param.F
                 / 3600,
             }