+++ tests/integration/test_models/standard_output_tests.py
@@ -293,8 +293,7 @@ class ParticleConcentrationTests(BaseOutputTest):
         self.N_s_n = solution[f"Negative {self.phase_name_n}particle flux"]
         self.N_s_p = solution[f"Positive {self.phase_name_p}particle flux"]
 
-        self.c_SEI_tot = solution["Loss of lithium to SEI [mol]"]
-        self.c_Li_tot = solution["Loss of lithium to lithium plating [mol]"]
+        self.c_Li_tot = solution["Total lithium in system [mol]"]
 
         if model.options["particle size"] == "distribution":
             # These concentration variables are only present for distribution models.
@@ -401,12 +400,7 @@ class ParticleConcentrationTests(BaseOutputTest):
     def test_conservation(self):
         """Test amount of lithium stored across all particles and in SEI layers is
         constant."""
-        c_s_tot = (
-            self.c_s_n_tot(self.solution.t)
-            + self.c_s_p_tot(self.solution.t)
-            + self.c_SEI_tot(self.solution.t)
-            + self.c_Li_tot(self.solution.t)
-        )
+        c_s_tot = self.c_Li_tot(self.solution.t)
         diff = (c_s_tot[1:] - c_s_tot[:-1]) / c_s_tot[:-1]
         if self.model.options["particle"] == "quartic profile":
             np.testing.assert_array_almost_equal(diff, 0, decimal=10)
@@ -800,20 +794,40 @@ class DegradationTests(BaseOutputTest):
         self.LLI = solution["Loss of lithium inventory [%]"]
         self.n_Li_lost = solution["Total lithium lost [mol]"]
         self.n_Li_lost_rxn = solution["Total lithium lost to side reactions [mol]"]
+        self.n_Li_lost_LAM = solution["Total lithium lost to LAM [mol]"]
+        self.n_Li_lost_LAM_n = solution[
+            "Loss of lithium due to loss of active material in negative electrode [mol]"
+        ]
+        self.n_Li_lost_LAM_p = solution[
+            "Loss of lithium due to loss of active material in positive electrode [mol]"
+        ]
 
     def test_degradation_modes(self):
         """Test degradation modes are between 0 and 100%"""
         np.testing.assert_array_less(-3e-3, self.LLI(self.t))
         np.testing.assert_array_less(-1e-13, self.LAM_ne(self.t))
         np.testing.assert_array_less(-1e-13, self.LAM_pe(self.t))
+        np.testing.assert_array_less(-1e-13, self.n_Li_lost_LAM_n(self.t))
+        np.testing.assert_array_less(-1e-13, self.n_Li_lost_LAM_p(self.t))
         np.testing.assert_array_less(self.LLI(self.t), 100)
         np.testing.assert_array_less(self.LAM_ne(self.t), 100)
         np.testing.assert_array_less(self.LAM_pe(self.t), 100)
 
     def test_lithium_lost(self):
         """Test the two ways of measuring lithium lost give the same value"""
+        print(
+            np.max(
+                np.abs(
+                    self.n_Li_lost(self.t)
+                    - self.n_Li_lost_rxn(self.t)
+                    - self.n_Li_lost_LAM(self.t)
+                )
+            )
+        )
         np.testing.assert_array_almost_equal(
-            self.n_Li_lost(self.t), self.n_Li_lost_rxn(self.t), decimal=3
+            self.n_Li_lost(self.t),
+            self.n_Li_lost_rxn(self.t) + self.n_Li_lost_LAM(self.t),
+            decimal=3,
         )
 
     def test_all(self):